feat: consolidate cloud observability into adapters/modules/parlant_cloud.py
Move ParlantCloudTracer, ParlantCloudLogger, ParlantCloudMeter and
configure_container() into a single file at adapters/modules/parlant_cloud.py.
Remove the adapters/observability/ directory. Server auto-loads the module
when PARLANT_CLOUD_API_KEY is set, reading project_id from the auth response.
